Retain issuer subpath when computing well-known configuration URI
In some cases, the issuer identifier might contain a subpath. This
subpath must be retained when doing discovery. This change keeps
the subpath if any to build the URL to the discovery document when
lico is setup to use an external OpenID Connect authority for its
identity provider.

Reference: https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderConfigurationRequest