Fix parsing of JWKS in authorities registration YAML
The autorities can have a local inline configuration for a JWKS
which is parsed using a struct which only has JSON unmarshalling
code. With this change, the YAML is first converted to JSON and
all the unmarshalling is handled by the corresponding JSON code
neatly fixing the issue.

Without this change, the inline JWKS keys are not parsed and fail
with `failed to decode public key` error message at startup.