Properly handle prompt select_account and consent for external oidc
The prompt values for select_account and consent need to be treated so
that there is no loop. This change ensures that select_account is only
handled by the external authority while consent is only handled by the
internal authority. Furthermore, any external callback which yields an
internal interactive identifier no longer always loops back to the
external authority when coming from the external callback.