Increase state cookie duration to 10 minutes
Sometimes it takes a while to enter username and password. The state
cookie expires rather quickly and if a user takes longer than that
expiration, the sign in fails. This change increases the duration of the
state cookie from 1 minute to 10 minutes which should be plenty for all
valid use cases.