Escape LDAP filter values when constructing filters
When constructing LDAP filters for looking up users by name or ID
make sure the assertion values are correctly escaped. Otherwise we might
be creating invalid filters e.g. when the username contains the '(',
')', '*' or '\' character.
Or when when using a binary attribute such as 'objectGUID' in
LDAP_SUB_ATTRIBUTES.