Fix branding settings cache usage
The branding settings are delivered as part of the initial hello and
then are cached in the state store. Any consumer must use the branding
settings from there instead from a potentially different hello state to
avoid loosing branding settings.

This fixes the flicker of the login form label and bottom text when the
login page is submitted.