Remove :443 from Host header for secure referrer/origin check
In certain cases the incoming request HTTP Host header might contain the
port (for example for HTTP/3 requests). This makes the host/referrer
check fail as those will not include this port. This change simply
always strips the port from the incoming Host header if it is 443 to
deal with the expected cases if the web server in from of the identifier
serves requests with HTTP3 on port 443.