Implement refresh and revoke for lg identifier backend session
With this change the libregraph identifier backend does have an
implementation for the refresh session and destroy session calls.

Refreshing a session simply loads the user from the backend, providing
the current session ID in the request headers similar to the already
existing request headers.

Destroying a session implements a new endpoint which is modeled after
the Microsoft Graph revokeSignInSessions but is only for a single
session (hooked at `/api/v1/me/revokeSignInSession`). If a backend
implements this endpoint it must do so for a POST request and return an
odata boolean value.