Keep extra backend provided id and access token claims on refresh
When using a refresh token, any potentially encoded extra claims for id
and access token must be retained. With this change, the claims embedded
into the refresh token are applied on top of any other access or id
token claims before creating the corresponding token. This avoids
loosing any of the backend provided claims when refresh tokens are used.