Allow client redirect URI with only a scheme
The redirect URI can have the form `protocol:action` which is basically
just a scheme. This is entirely valid and since such URI do not have a
host the validation of the redirect_uri request parameter fails.

This change does no longer check the host part of the URI to be
non-empty, fixing client support for clients using such URIs.

Fixes: #161